fixed some shit

+++ laba_1/src/group5/varC/sushko/products/Components/Bread.java
@@ -3,7 +3,7 @@ package group5.varC.sushko.products.Components;
 import group5.varC.sushko.products.baseClass.Components;
 
 public class Bread extends Components {
-    public Bread(double mass){
+    public Bread(int mass){
         this.mass = mass;
     }
     static double caloriesPerGram = 2.65;
