+++ laba_1/src/group5/varC/sushko/products/Components/Milk.java
@@ -3,7 +3,7 @@ package group5.varC.sushko.products.Components;
 import group5.varC.sushko.products.baseClass.Components;
 
 public class Milk extends Components {
-    public Milk(double mass){
+    public Milk(int mass){
         this.mass = mass;
     }
     static double caloriesPerGram = 0.42;
