+++ laba_1/src/group5/varC/sushko/products/Components/Potato.java
@@ -3,7 +3,7 @@ package group5.varC.sushko.products.Components;
 import group5.varC.sushko.products.baseClass.Components;
 
 public class Potato extends Components {
-    public Potato(double mass){
+    public Potato(int mass){
         this.mass = mass;
     }
     @Override
