+++ laba_1/src/group5/varC/sushko/products/baseClass/Products.java
@@ -3,18 +3,18 @@ package group5.varC.sushko.products.baseClass;
 import group5.varC.sushko.products.interfaces.Nutritious;
 
 public abstract class Products implements Nutritious {
-    protected double mass;
-    protected double price;
-    public double getPrice() {
+    protected int mass;
+    protected int price;
+    public int getPrice() {
         return price;
     }
-    public double getMass() {
+    public int getMass() {
         return mass;
     }
-    public void setMass(double mass) {
+    public void setMass(int mass) {
         this.mass = mass;
     }
-    public void setPrice(double price) {
+    public void setPrice(int price) {
         this.price = price;
     }
